A002071: Split Q by p
Each prime now only considers the Q which involve p.
Extra work is done at each p to consider p-smooth up to P BUT this avoid
redoing each Q |primes| times!

Because Q is 2^p_i this is very close to a 2x speedup!

In practice smaller Q have to consider up to a higher `MAX_CF` so more
smaller Q are expanded but that's a good double check and doesn't matter
very much because they were always expanded up to the max(`MAX_CF`) for
the final p.